refactor(fix-wizard): fix_wizard_dialog.py(1857줄)를 5단계 페이지로 분리
대형 파일 정리 3번째(Phase C). QWizard 각 단계 페이지가 서로 직접 참조하지
않고 wizard_dialog(글루) 매개로만 상태를 공유하는 구조라 db_dialogs.py와
동일한 파사드 패턴으로 안전하게 분리 가능했음.

- fix_wizard_issue_selection_page.py (신규): IssueSelectionPage
- fix_wizard_charset_page.py (신규): CharsetFixPage
- fix_wizard_option_page.py (신규): FixOptionPage + BatchOptionDialog
- fix_wizard_preview_page.py (신규): PreviewPage
- fix_wizard_execution_page.py (신규): ExecutionPage
- fix_wizard_dialog.py: 1857 -> 156줄 (FixWizardDialog + 전체 재수입)

각 페이지의 `wizard: FixWizardDialog` 타입힌트는 순환 import를 유발해
`wizard: "FixWizardDialog"` 문자열 forward-reference로 변경(런타임 미평가).

테스트 수정: ExecutionPage.execute()가 호출하는 FixWizardWorker는 이제
fix_wizard_execution_page 모듈 네임스페이스에서 resolve되므로, 해당
monkeypatch 대상을 fix_wizard_dialog -> fix_wizard_execution_page로 이동.
QMessageBox.warning 패치는 공유 Qt 클래스 메서드 패치라 그대로 유효.

검증: 전체 pytest 1809 passed / 1 known-flaky 제외. py_compile 통과.
FixWizardDialog 오프스크린 오픈 확인. 앱 스모크(core_hello) 성공.
